requirements.txt
pygame>=2.0.1
numpy>=1.18.5
pandas>=1.0.5
Pillow>=7.2.0
ffmpeg-python>=0.2.0
pydub>=0.25.1
openpyxl>=3.0.4
azure-cognitiveservices-speech>=1.21.0
ttkbootstrap>=1.10.0
tkextrafont==0.6.3
chlorophyll==0.3.1